refactor: 拆分会话存储模块
