+++ qq-maid-core/src/storage/session/mod.rs
@@ -4,31 +4,33 @@
 //! 原有的整条会话读写接口。业务层仍然操作 `SessionRecord`，存储层负责把
 //! 会话元信息、活跃会话映射和消息顺序拆分保存到数据库中。
 
-use std::fmt;
-
 pub use qq_maid_common::redaction::redact_sensitive_text;
 use qq_maid_common::time_context;
-use rusqlite::{Connection, OptionalExtension, Row, Transaction, params};
-use serde::{Deserialize, Serialize};
+use rusqlite::{Connection, OptionalExtension, params};
 use serde_json::{Map, Value};
 
-use crate::{
-    identity::{parse_stable_scope_key, stable_scope_key},
-    runtime::pending::PendingOperation,
-    runtime::tools::todo::{TodoItem, TodoStatus},
-    storage::database::{DatabaseError, SqliteDatabase, SqliteMigration},
-};
+use crate::storage::database::{SqliteDatabase, SqliteMigration};
 
+mod error;
 mod jsonio;
+mod model;
 mod normalize;
 mod row;
+mod write;
 
-use jsonio::{encode_json, encode_optional_json};
+pub use error::SessionError;
+pub use model::{
+    LastMemoryQuery, LastTodoAction, LastTodoQuery, SessionMessage, SessionMeta, SessionRecord,
+};
 use normalize::{
     build_session_id, infer_scope, initial_session_state, normalize_session,
     normalize_session_title,
 };
 use row::load_session_unlocked;
+use write::{
+    collect_sql_rows, replace_messages_tx, set_active_session_id_conn, set_active_session_id_tx,
+    upsert_session_tx,
+};
 
 /// Session schema migration，由应用启动时的通用数据库初始化流程统一执行。
 ///
@@ -119,116 +121,6 @@ pub const DEFAULT_SESSION_TITLE: &str = "未命名会话";
 /// 最近列表/查询快照的统一有效期（秒）。
 pub const LAST_QUERY_TTL_SECONDS: i64 = 10 * 60;
 
-/// 会话记录，包含完整的会话状态和历史。
-///
-/// 每个会话对应一个 scope_key（如 "group:g1"），包含消息历史、对话摘要、
-/// 对话状态、挂起操作以及上次的待办/记忆查询记录。
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct SessionRecord {
-    #[serde(default)]
-    pub session_id: String,
-    #[serde(default)]
-    pub scope: String,
-    #[serde(default)]
-    pub scope_key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub user_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub group_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub guild_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub channel_id: Option<String>,
-    #[serde(default)]
-    pub platform: String,
-    #[serde(default)]
-    pub created_at: String,
-    #[serde(default)]
-    pub updated_at: String,
-    #[serde(default)]
-    pub title: String,
-    #[serde(default)]
-    pub state: Map<String, Value>,
-    #[serde(default)]
-    pub summary: String,
-    #[serde(default)]
-    pub history: Vec<SessionMessage>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub pending_operation: Option<PendingOperation>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub last_todo_query: Option<LastTodoQuery>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub last_todo_action: Option<LastTodoAction>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub last_memory_query: Option<LastMemoryQuery>,
-    #[serde(default, flatten)]
-    pub extra: Map<String, Value>,
-}
-
-/// 会话中的单条消息，包含角色、内容和时间戳。
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct SessionMessage {
-    pub role: String,
-    pub content: String,
-    pub ts: String,
-}
-
-/// 上次待办查询记录，用于在会话上下文中快速引用查询结果。
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct LastTodoQuery {
-    pub owner_key: String,
-    pub query_type: String,
-    pub condition: String,
-    #[serde(default)]
-    pub result_ids: Vec<String>,
-    pub created_at: String,
-}
-
-/// 最近一次成功改变 Todo 状态的条目快照。
-///
-/// 该结构只保存“刚才那个/它/恢复的那个”所需的最小信息；真正执行新操作时，
-/// 仍必须回到 TodoStore 用 owner + item_id 再查一次当前状态，不能信任 session 缓存。
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct LastTodoAction {
-    pub owner_key: String,
-    pub item_id: String,
-    pub title: String,
-    pub action: String,
-    pub resulting_status: TodoStatus,
-    pub created_at: String,
-}
-
-/// 上次记忆查询记录，用于在会话上下文中快速引用查询结果。
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct LastMemoryQuery {
-    pub query_type: String,
-    pub condition: String,
-    /// 列表生成时的记忆访问边界；旧快照缺失时运行时会要求重新列表。
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub scope_type: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub scope_id: Option<String>,
-    #[serde(default)]
-    pub result_ids: Vec<String>,
-    pub created_at: String,
-}
-
-/// 会话元信息，用于标识和创建会话。
-///
-/// 包含作用域、作用域键值、用户/群组/频道信息以及平台标识。
-/// scope_key 的格式如 "group:g1"、"private:u1"、"guild:guild_id:channel_id"。
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SessionMeta {
-    pub scope: String,
-    pub scope_key: String,
-    pub user_id: Option<String>,
-    pub group_id: Option<String>,
-    pub guild_id: Option<String>,
-    pub channel_id: Option<String>,
-    pub platform: String,
-    pub account_id: Option<String>,
-}
-
 /// 会话存储器，基于项目通用 SQLite 连接实现。
 ///
 /// 数据库连接由应用启动时统一打开并执行 migration；SessionStore 不再读取
@@ -236,13 +128,6 @@ pub struct SessionMeta {
 #[derive(Debug, Clone)]
 pub struct SessionStore {
     database: SqliteDatabase,
-}
-
-/// 会话操作错误类型。
-#[derive(Debug, Clone)]
-pub struct SessionError {
-    code: &'static str,
-    message: String,
 }
 
 impl SessionStore {
@@ -518,388 +403,6 @@ impl SessionStore {
         replace_messages_tx(&tx, session)?;
         tx.commit().map_err(SessionError::from_sql)
     }
-}
-
-impl SessionRecord {
-    /// 追加一条消息到会话历史（仅允许 user 和 assistant 角色）。
-    /// 内容会自动脱敏。
-    pub fn append_message(&mut self, role: &str, content: &str) {
-        if !matches!(role, "user" | "assistant") {
-            return;
-        }
-        self.history.push(SessionMessage {
-            role: role.to_owned(),
-            content: redact_sensitive_text(content),
-            ts: now_iso_cn(),
-        });
-    }
-
-    /// 重置会话上下文：清空历史、摘要、状态和挂起操作，保留会话元信息。
-    pub fn reset(&mut self) {
-        self.summary.clear();
-        self.state.clear();
-        self.history.clear();
-        self.pending_operation = None;
-        self.last_todo_query = None;
-        self.last_todo_action = None;
-        self.last_memory_query = None;
-    }
-
-    /// 合并追加回复前已由业务 flow 更新的短期交互状态。
-    ///
-    /// `append_exchange_with_latest` 会重新读取数据库中的 latest session；调用方手里的
-    /// current 可能已经更新 pending、最近查询或最近操作快照。这里集中维护这些
-    /// 运行态字段，避免最新数据库行反向覆盖同轮业务副作用。
-    pub fn merge_interaction_side_effects_from(&mut self, current: &SessionRecord) {
-        self.state = current.state.clone();
-        self.pending_operation = current.pending_operation.clone();
-        self.last_memory_query = current.last_memory_query.clone();
-        self.last_todo_query = current.last_todo_query.clone();
-        self.last_todo_action = current.last_todo_action.clone();
-    }
-
-    /// 记录最近一次真正展示给用户的 Todo 列表快照。
-    ///
-    /// `result_ids` 必须与最终展示顺序完全一致；后续“第一条 / 第二条 / 它”
-    /// 等续指只允许按这份快照映射，不能回退数据库默认顺序。
-    pub fn remember_last_todo_query(
-        &mut self,
-        owner_key: &str,
-        query_type: impl Into<String>,
-        condition: impl Into<String>,
-        result_ids: Vec<String>,
-    ) {
-        self.last_todo_query = Some(LastTodoQuery {
-            owner_key: owner_key.to_owned(),
-            query_type: query_type.into(),
-            condition: condition.into(),
-            result_ids,
-            created_at: now_iso_cn(),
-        });
-    }
-
-    /// 记录最近一次成功操作的单条 Todo。
-    ///
-    /// 这里只保存自然语言续指所需的最小快照；下次真正执行时仍需重新读取当前 Todo。
-    pub fn remember_last_todo_action(&mut self, owner_key: &str, item: &TodoItem, action: &str) {
-        self.last_todo_action = Some(LastTodoAction {
-            owner_key: owner_key.to_owned(),
-            item_id: item.id.clone(),
-            title: item.title.clone(),
-            action: action.to_owned(),
-            resulting_status: item.status.clone(),
-            created_at: if item.updated_at.trim().is_empty() {
-                now_iso_cn()
-            } else {
-                item.updated_at.clone()
-            },
-        });
-    }
-
-    /// 根据一次批量结果维护最近操作对象。
-    ///
-    /// 成功 0 条时保持原值，成功 1 条时记录该条，成功多条时清空，避免“刚才那个”歧义。
-    pub fn update_last_todo_action_from_items(
-        &mut self,
-        owner_key: &str,
-        action: &str,
-        items: &[TodoItem],
-    ) {
-        match items {
-            [] => {}
-            [item] => self.remember_last_todo_action(owner_key, item, action),
-            _ => self.last_todo_action = None,
-        }
-    }
-
-    /// 当物理删除命中最近对象时清空该快照，避免 session 持续引用已不存在条目。
-    pub fn clear_last_todo_action_if_matches_any(&mut self, owner_key: &str, item_ids: &[String]) {
-        let should_clear = self.last_todo_action.as_ref().is_some_and(|last_action| {
-            last_action.owner_key == owner_key
-                && item_ids
-                    .iter()
-                    .any(|item_id| item_id == &last_action.item_id)
-        });
-        if should_clear {
-            self.last_todo_action = None;
-        }
-    }
-}
-
-impl SessionMeta {
-    /// 创建会话元信息，自动推断作用域类型（guild_channel / group / private）。
-    pub fn new(
-        scope_key: impl Into<String>,
-        user_id: Option<String>,
-        group_id: Option<String>,
-        guild_id: Option<String>,
-        channel_id: Option<String>,
-        platform: impl Into<String>,
-    ) -> Self {
-        Self::new_with_account(
-            scope_key, user_id, group_id, guild_id, channel_id, platform, None,
-        )
-    }
-
-    /// 创建带平台账号维度的会话元信息。
-    ///
-    /// account_id 只用于业务隔离键和后续 owner/scope 推导，不是平台发送目标。
-    pub fn new_with_account(
-        scope_key: impl Into<String>,
-        user_id: Option<String>,
-        group_id: Option<String>,
-        guild_id: Option<String>,
-        channel_id: Option<String>,
-        platform: impl Into<String>,
-        account_id: Option<String>,
-    ) -> Self {
-        let scope_key = scope_key.into();
-        let scope = infer_scope(&scope_key, group_id.as_deref(), guild_id.as_deref());
-        Self {
-            scope,
-            scope_key,
-            user_id,
-            group_id,
-            guild_id,
-            channel_id,
-            platform: platform.into(),
-            account_id,
-        }
-    }
-
-    /// 当前 actor 的个人业务隔离键。
-    ///
-    /// 返回值用于 Memory / Todo 等业务归属判断；平台发送仍使用原始 user_id。
-    pub fn personal_scope_id(&self) -> Option<String> {
-        let user_id = clean_optional_str(self.user_id.as_deref())?;
-        if should_namespace_scope(self) {
-            Some(stable_scope_key(
-                platform_or_default(&self.platform),
-                self.account_id.as_deref(),
-                "private",
-                user_id,
-            ))
-        } else {
-            Some(user_id.to_owned())
-        }
-    }
-
-    /// 当前群会话的群级业务隔离键。
-    ///
-    /// 返回值只用于群 Memory / 群 Pending 等状态隔离，不作为群消息发送目标。
-    pub fn group_scope_id(&self) -> Option<String> {
-        let group_id = clean_optional_str(self.group_id.as_deref())?;
-        if let Some(parsed) = parse_stable_scope_key(&self.scope_key)
-            && parsed.target_type == "group"
-        {
-            return Some(self.scope_key.clone());
-        }
-        if should_namespace_scope(self) {
-            Some(stable_scope_key(
-                platform_or_default(&self.platform),
-                self.account_id.as_deref(),
-                "group",
-                group_id,
-            ))
-        } else {
-            Some(group_id.to_owned())
-        }
-    }
-}
-
-fn should_namespace_scope(meta: &SessionMeta) -> bool {
-    meta.account_id
-        .as_deref()
-        .is_some_and(|value| !value.trim().is_empty())
-        || parse_stable_scope_key(&meta.scope_key).is_some()
-}
-
-fn platform_or_default(value: &str) -> &str {
-    let value = value.trim();
-    if value.is_empty() { "qq" } else { value }
-}
-
-fn clean_optional_str(value: Option<&str>) -> Option<&str> {
-    value.map(str::trim).filter(|value| !value.is_empty())
-}
-
-impl SessionError {
-    /// 获取错误码。
-    pub fn code(&self) -> &str {
-        self.code
-    }
-
-    /// 获取错误消息。
-    pub fn message(&self) -> &str {
-        &self.message
-    }
-
-    fn encode(message: impl Into<String>) -> Self {
-        Self {
-            code: "encode_error",
-            message: message.into(),
-        }
-    }
-
-    fn decode(message: impl Into<String>) -> Self {
-        Self {
-            code: "decode_error",
-            message: message.into(),
-        }
-    }
-
-    fn data(message: impl Into<String>) -> Self {
-        Self {
-            code: "data_error",
-            message: message.into(),
-        }
-    }
-
-    fn from_database(err: DatabaseError) -> Self {
-        Self {
-            code: "database_error",
-            message: format!("sqlite database failed: {}", err.message()),
-        }
-    }
-
-    fn from_sql(err: rusqlite::Error) -> Self {
-        Self {
-            code: "database_error",
-            message: format!("sqlite session failed: {err}"),
-        }
-    }
-}
-
-impl fmt::Display for SessionError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}: {}", self.code, self.message)
-    }
-}
-
-impl std::error::Error for SessionError {}
-
-fn upsert_session_tx(tx: &Transaction<'_>, session: &SessionRecord) -> Result<(), SessionError> {
-    let state_json = encode_json(&session.state, "session state")?;
-    let pending_operation_json = encode_optional_json(&session.pending_operation, "pending")?;
-    let last_todo_query_json = encode_optional_json(&session.last_todo_query, "last todo query")?;
-    let last_todo_action_json =
-        encode_optional_json(&session.last_todo_action, "last todo action")?;
-    let last_memory_query_json =
-        encode_optional_json(&session.last_memory_query, "last memory query")?;
-    let extra_json = encode_json(&session.extra, "session extra")?;
-    tx.execute(
-        "INSERT INTO sessions (
-            session_id, scope, scope_key, user_id, group_id, guild_id, channel_id, platform,
-            created_at, updated_at, title, state_json, summary, pending_operation_json,
-            last_todo_query_json, last_todo_action_json, last_memory_query_json, extra_json
-         ) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14, ?15, ?16, ?17, ?18)
-         ON CONFLICT(session_id) DO UPDATE SET
-            scope = excluded.scope,
-            scope_key = excluded.scope_key,
-            user_id = excluded.user_id,
-            group_id = excluded.group_id,
-            guild_id = excluded.guild_id,
-            channel_id = excluded.channel_id,
-            platform = excluded.platform,
-            created_at = excluded.created_at,
-            updated_at = excluded.updated_at,
-            title = excluded.title,
-            state_json = excluded.state_json,
-            summary = excluded.summary,
-            pending_operation_json = excluded.pending_operation_json,
-            last_todo_query_json = excluded.last_todo_query_json,
-            last_todo_action_json = excluded.last_todo_action_json,
-            last_memory_query_json = excluded.last_memory_query_json,
-            extra_json = excluded.extra_json",
-        params![
-            session.session_id.as_str(),
-            session.scope.as_str(),
-            session.scope_key.as_str(),
-            session.user_id.as_deref(),
-            session.group_id.as_deref(),
-            session.guild_id.as_deref(),
-            session.channel_id.as_deref(),
-            session.platform.as_str(),
-            session.created_at.as_str(),
-            session.updated_at.as_str(),
-            session.title.as_str(),
-            state_json,
-            session.summary.as_str(),
-            pending_operation_json,
-            last_todo_query_json,
-            last_todo_action_json,
-            last_memory_query_json,
-            extra_json,
-        ],
-    )
-    .map_err(SessionError::from_sql)?;
-    Ok(())
-}
-
-fn replace_messages_tx(tx: &Transaction<'_>, session: &SessionRecord) -> Result<(), SessionError> {
-    tx.execute(
-        "DELETE FROM session_messages WHERE session_id = ?1",
-        params![session.session_id.as_str()],
-    )
-    .map_err(SessionError::from_sql)?;
-    for (index, message) in session.history.iter().enumerate() {
-        tx.execute(
-            "INSERT INTO session_messages (session_id, message_index, role, content, ts)
-             VALUES (?1, ?2, ?3, ?4, ?5)",
-            params![
-                session.session_id.as_str(),
-                index as i64,
-                message.role.as_str(),
-                message.content.as_str(),
-                message.ts.as_str(),
-            ],
-        )
-        .map_err(SessionError::from_sql)?;
-    }
-    Ok(())
-}
-
-fn set_active_session_id_conn(
-    conn: &Connection,
-    scope_key: &str,
-    session_id: &str,
-    now: &str,
-) -> Result<(), SessionError> {
-    conn.execute(
-        "INSERT INTO session_active (scope_key, session_id, updated_at)
-         VALUES (?1, ?2, ?3)
-         ON CONFLICT(scope_key) DO UPDATE SET
-            session_id = excluded.session_id,
-            updated_at = excluded.updated_at",
-        params![scope_key, session_id, now],
-    )
-    .map_err(SessionError::from_sql)?;
-    Ok(())
-}
-
-fn set_active_session_id_tx(
-    tx: &Transaction<'_>,
-    scope_key: &str,
-    session_id: &str,
-    now: &str,
-) -> Result<(), SessionError> {
-    tx.execute(
-        "INSERT INTO session_active (scope_key, session_id, updated_at)
-         VALUES (?1, ?2, ?3)
-         ON CONFLICT(scope_key) DO UPDATE SET
-            session_id = excluded.session_id,
-            updated_at = excluded.updated_at",
-        params![scope_key, session_id, now],
-    )
-    .map_err(SessionError::from_sql)?;
-    Ok(())
-}
-fn collect_sql_rows<T, F>(rows: rusqlite::MappedRows<'_, F>) -> Result<Vec<T>, SessionError>
-where
-    F: FnMut(&Row<'_>) -> rusqlite::Result<T>,
-{
-    rows.collect::<Result<Vec<_>, _>>()
-        .map_err(SessionError::from_sql)
 }
 
 /// 获取当前北京时间 ISO8601 字符串。
